--- pkg/common/templates.go
@@ -1,62 +1,0 @@
-package common
-
-import (
-	"fmt"
-	"io/ioutil"
-	"os"
-
-	"github.com/flyteorg/flyteadmin/pkg/errors"
-	"github.com/flyteorg/flyteadmin/pkg/runtime/interfaces"
-	"google.golang.org/grpc/codes"
-)
-
-const NamespaceVariable = "namespace"
-const ProjectVariable = "project"
-const DomainVariable = "domain"
-const TemplateVariableFormat = "{{ %s }}"
-
-type TemplateValuesType = map[string]string
-
-// Given a map of templatized variable names -> data source, this function produces an output that maps the same
-// variable names to their fully resolved values (from the specified data source).
-func PopulateTemplateValues(data map[string]interfaces.DataSource) (TemplateValuesType, error) {
-	templateValues := make(TemplateValuesType, len(data))
-	collectedErrs := make([]error, 0)
-	for templateVar, dataSource := range data {
-		if templateVar == NamespaceVariable || templateVar == ProjectVariable || templateVar == DomainVariable {
-			// The namespace variable is specifically reserved for system use only.
-			collectedErrs = append(collectedErrs, errors.NewFlyteAdminErrorf(codes.InvalidArgument,
-				"Cannot assign namespace template value in user data"))
-			continue
-		}
-		var dataValue string
-		if len(dataSource.Value) > 0 {
-			dataValue = dataSource.Value
-		} else if len(dataSource.ValueFrom.EnvVar) > 0 {
-			dataValue = os.Getenv(dataSource.ValueFrom.EnvVar)
-		} else if len(dataSource.ValueFrom.FilePath) > 0 {
-			templateFile, err := ioutil.ReadFile(dataSource.ValueFrom.FilePath)
-			if err != nil {
-				collectedErrs = append(collectedErrs, errors.NewFlyteAdminErrorf(codes.InvalidArgument,
-					"failed to substitute parameterized value for %s: unable to read value from: [%+v] with err: %v",
-					templateVar, dataSource.ValueFrom.FilePath, err))
-				continue
-			}
-			dataValue = string(templateFile)
-		} else {
-			collectedErrs = append(collectedErrs, errors.NewFlyteAdminErrorf(codes.InvalidArgument,
-				"failed to substitute parameterized value for %s: unset or unrecognized ValueFrom: [%+v]", templateVar, dataSource.ValueFrom))
-			continue
-		}
-		if len(dataValue) == 0 {
-			collectedErrs = append(collectedErrs, errors.NewFlyteAdminErrorf(codes.InvalidArgument,
-				"failed to substitute parameterized value for %s: unset. ValueFrom: [%+v]", templateVar, dataSource.ValueFrom))
-			continue
-		}
-		templateValues[fmt.Sprintf(TemplateVariableFormat, templateVar)] = dataValue
-	}
-	if len(collectedErrs) > 0 {
-		return nil, errors.NewCollectedFlyteAdminError(codes.InvalidArgument, collectedErrs)
-	}
-	return templateValues, nil
-}
